requirements.txt
pandas==2.1.1
numpy==1.26.0
statsmodels==0.14.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
matplotlib
seaborn
statsmodels
pandas
sqlalchemy